refactor: Collapse nested test layout to sibling tests/ for orbit-exec … [ORB-00241]
Collapse nested test layout to sibling tests/ for orbit-exec and orbit-knowledge
