+++ crates/orbit-exec/src/macos_sandbox/sbpl_filter.rs
@@ -108,5 +108,3 @@ fn subpath_root(rule: &str) -> String {
     }
 }
 
-#[cfg(test)]
-mod tests;
