+++ crates/orbit-exec/src/macos_sandbox/tests/mod.rs
@@ -1,0 +1,6 @@
+#![allow(missing_docs)]
+
+mod compile;
+mod provider_dirs;
+mod sbpl_filter;
+mod spawn;
