+++ crates/orbit-exec/src/macos_sandbox/tests/spawn.rs
@@ -2,7 +2,7 @@
 use super::super::test_support::{
     ScopeGuard, sandbox_exec_can_apply, sandbox_test_parent, shell_escape,
 };
-use super::{MacosSandboxSpawnRequest, sandbox_exec_path_from, spawn_under_macos_sandbox};
+use super::super::spawn::{MacosSandboxSpawnRequest, sandbox_exec_path_from, spawn_under_macos_sandbox};
 use std::path::Path;
 use std::process::Stdio;
