+++ crates/orbit-knowledge/src/graph/sqlite_index/reader.rs
@@ -439,5 +439,3 @@ impl GraphIndexReader {
     }
 }
 
-#[cfg(test)]
-mod tests;
