--- crates/orbit-knowledge/src/graph/sqlite_index/reader/tests/mod.rs
@@ -1,6 +1,0 @@
-//! Unit tests for `reader`, split by concern per test_layout.md.
-//!
-//! All tests here exercise GraphIndexReader construction / open paths
-//! (missing, stale, valid cases, read-only enforcement).
-
-mod open;
