--- crates/orbit-knowledge/src/pipeline/scan/tests/mod.rs
@@ -1,3 +1,0 @@
-#![allow(missing_docs)]
-
-mod scan;
