+++ crates/orbit-knowledge/src/pipeline/tests/mod.rs
@@ -1,3 +1,5 @@
 #![allow(missing_docs)]
 
+mod hash;
 mod integration;
+mod scan;
